Move types to separate files.
