+++ src/Microsoft.Tye.Extensions/Dapr/DaprExtension.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
@@ -13,71 +12,6 @@ using System.Threading.Tasks;
 
 namespace Microsoft.Tye.Extensions.Dapr
 {
-    internal abstract class DaprExtensionCommonConfiguration
-    {
-        public int? PlacementPort  { get; set; }
-    }
-
-    internal sealed class DaprExtensionServiceConfiguration : DaprExtensionCommonConfiguration
-    {
-        public bool? Enabled { get; set; }
-    }
-
-    internal sealed class DaprExtensionConfiguration : DaprExtensionCommonConfiguration
-    {
-        public IReadOnlyDictionary<string, DaprExtensionServiceConfiguration> Services { get; set; }
-            = new Dictionary<string, DaprExtensionServiceConfiguration>();
-    }
-
-    internal static class DaprExtensionConfigurationReader
-    {
-        public static DaprExtensionConfiguration ReadConfiguration(IDictionary<string, object> rawConfiguration)
-        {
-            var configuration = new DaprExtensionConfiguration();
-
-            ReadCommonConfiguration(rawConfiguration, configuration);
-
-            if (rawConfiguration.TryGetValue("services", out var servicesObject) && servicesObject is Dictionary<string, object> rawServicesConfiguration)
-            {
-                var services = new Dictionary<string, DaprExtensionServiceConfiguration>();
-
-                foreach (var kvp in rawServicesConfiguration)
-                {
-                    if (kvp.Value is Dictionary<string, object> rawServiceConfiguration)
-                    {
-                        var serviceConfiguration = new DaprExtensionServiceConfiguration();
-
-                        ReadServiceConfiguration(rawServiceConfiguration, serviceConfiguration);
-
-                        services.Add(kvp.Key, serviceConfiguration);                        
-                    }
-                }
-
-                configuration.Services = services;
-            }
-
-            return configuration;
-        }
-
-        private static void ReadServiceConfiguration(IDictionary<string, object> rawConfiguration, DaprExtensionServiceConfiguration serviceConfiguration)
-        {
-            ReadCommonConfiguration(rawConfiguration, serviceConfiguration);
-
-            if (rawConfiguration.TryGetValue("enabled", out var obj) && obj is string && Boolean.TryParse(obj.ToString(), out var enabled))
-            {
-                serviceConfiguration.Enabled = enabled;
-            }
-        }
-
-        private static void ReadCommonConfiguration(IDictionary<string, object> rawConfiguration, DaprExtensionCommonConfiguration commonConfiguration)
-        {
-            if (rawConfiguration.TryGetValue("placement-port", out var obj) && obj?.ToString() is string && int.TryParse(obj.ToString(), out var customPlacementPort))
-            {
-                commonConfiguration.PlacementPort = customPlacementPort;
-            }
-        }
-    }
-
     internal sealed class DaprExtension : Extension
     {
         public override async Task ProcessAsync(ExtensionContext context, ExtensionConfiguration config)
